+++ Cargo.toml
@@ -1,11 +1,13 @@
-[package]
-name = "inel"
-version = "0.1.0"
-edition = "2021"
+[workspace]
+resolver = "2"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+members = [
+  "crates/inel",
+  "crates/inel-executor",
+  "crates/inel-reactor",
+]
 
-[dependencies]
+[workspace.dependencies]
 flume = "0.11"
 futures = "0.3"
 io-uring = "0.6"
@@ -13,7 +15,3 @@ libc = "0.2"
 oneshot = "0.1"
 tracing = "0.1"
 pin-project-lite = "0.2"
-
-[dev-dependencies]
-tracing-subscriber = { version = "0.3", features = ["env-filter"] }
-rand = "0.8"
