+++ crates/inel-executor/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "inel-executor"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+flume = { workspace = true }
+futures = { workspace = true }
+tracing = { workspace = true }
+
+[dev-dependencies]
+tracing-subscriber = "0.3.18"
