+++ crates/inel-reactor/Cargo.toml
@@ -1,0 +1,4 @@
+[package]
+name = "inel-reactor"
+version = "0.1.0"
+edition = "2021"
