+++ crates/inel/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "inel"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+inel-executor = { path = "../inel-executor" }
