--- src/io/mod.rs
@@ -1,5 +1,0 @@
-pub mod unix {
-    pub use crate::reactor::socket::{bind, connect, listen, socket};
-}
-
-pub use crate::reactor::read::RingBufReader;
