--- src/lib.rs
@@ -1,16 +1,0 @@
-pub(crate) mod reactor;
-pub(crate) mod runtime;
-
-pub mod io;
-pub mod net;
-pub mod time;
-
-pub use runtime::{
-    executor::{block_on, run, spawn},
-    join::JoinHandle,
-};
-
-pub use reactor::{
-    read::{AsyncRingRead, RingRead},
-    write::{AsyncRingWrite, AsyncRingWriteExt, RingWrite},
-};
