--- tests/runtime/main.rs
@@ -1,5 +1,0 @@
-pub mod cancel;
-pub mod helpers;
-pub mod net;
-pub mod sanity;
-pub mod time;
